extract MethodRefLinker since this is a common util
