+++ joern-cli/frontends/x2cpg/src/main/scala/io/joern/x2cpg/passes/base/AstLinkerPass.scala
@@ -1,11 +1,6 @@
 package io.joern.x2cpg.passes.base
 
-import io.joern.x2cpg.passes.callgraph.MethodRefLinker
-import io.joern.x2cpg.passes.callgraph.MethodRefLinker.{
-  methodFullNameToNode,
-  namespaceBlockFullNameToNode,
-  typeDeclFullNameToNode
-}
+import io.joern.x2cpg.passes.EdgeLinker._
 import io.shiftleft.codepropertygraph.Cpg
 import io.shiftleft.codepropertygraph.generated.nodes.StoredNode
 import io.shiftleft.codepropertygraph.generated.{EdgeTypes, NodeTypes}
@@ -13,8 +8,6 @@ import io.shiftleft.passes.SimpleCpgPass
 import io.shiftleft.semanticcpg.language._
 
 class AstLinkerPass(cpg: Cpg) extends SimpleCpgPass(cpg) {
-
-  import MethodRefLinker.{logFailedSrcLookup, logger}
 
   override def run(dstGraph: DiffGraphBuilder): Unit = {
     cpg.method.whereNot(_.astParent).foreach { method =>
