+++ joern-cli/frontends/x2cpg/src/main/scala/io/joern/x2cpg/passes/typerelations/AliasLinkerPass.scala
@@ -1,10 +1,10 @@
 package io.joern.x2cpg.passes.typerelations
 
+import io.joern.x2cpg.passes.EdgeLinker.{linkToMultiple, typeFullNameToNode}
 import io.shiftleft.codepropertygraph.Cpg
 import io.shiftleft.codepropertygraph.generated.nodes.TypeDecl
 import io.shiftleft.codepropertygraph.generated.{EdgeTypes, NodeTypes, PropertyNames}
 import io.shiftleft.passes.SimpleCpgPass
-import io.joern.x2cpg.passes.callgraph.MethodRefLinker.{linkToMultiple, typeFullNameToNode}
 
 class AliasLinkerPass(cpg: Cpg) extends SimpleCpgPass(cpg) {
   override def run(dstGraph: DiffGraphBuilder): Unit = {
